+++ tests/streamlit/test_streamlit_manager.py
@@ -18,12 +18,13 @@ mock_streamlit_exists = mock.patch(
 )
 
 
-@mock.patch("snowflake.cli.api.artifacts.upload.StageManager")
+@mock.patch("snowflake.cli.api.entities.utils.StageManager")
+@mock.patch("snowflake.cli._plugins.stage.diff.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.get_url")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.execute_query")
 @mock_streamlit_exists
-def test_deploy_streamlit(mock_execute_query, _, mock_stage_manager, __):
+def test_deploy_streamlit(mock_execute_query, _, mock_stage_manager, __, ___):
     mock_stage_manager().get_standard_stage_prefix.return_value = "stage_root"
 
     main_file = Path("main.py")
@@ -59,13 +60,14 @@ def test_deploy_streamlit(mock_execute_query, _, mock_stage_manager, __):
     )
 
 
-@mock.patch("snowflake.cli.api.artifacts.upload.StageManager")
+@mock.patch("snowflake.cli.api.entities.utils.StageManager")
+@mock.patch("snowflake.cli._plugins.stage.diff.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.get_url")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.execute_query")
 @mock_streamlit_exists
 def test_deploy_streamlit_with_api_integrations(
-    mock_execute_query, _, mock_stage_manager, __
+    mock_execute_query, _, mock_stage_manager, __, ___
 ):
     mock_stage_manager().get_standard_stage_prefix.return_value = "stage_root"
 
@@ -104,13 +106,14 @@ def test_deploy_streamlit_with_api_integrations(
     )
 
 
-@mock.patch("snowflake.cli.api.artifacts.upload.StageManager")
+@mock.patch("snowflake.cli.api.entities.utils.StageManager")
+@mock.patch("snowflake.cli._plugins.stage.diff.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.get_url")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.execute_query")
 @mock_streamlit_exists
 def test_deploy_streamlit_with_comment(
-    mock_execute_query, _, mock_stage_manager, __, temporary_directory
+    mock_execute_query, _, mock_stage_manager, __, ___, temporary_directory
 ):
     mock_stage_manager().get_standard_stage_prefix.return_value = "stage_root"
 
@@ -146,13 +149,14 @@ def test_deploy_streamlit_with_comment(
     )
 
 
-@mock.patch("snowflake.cli.api.artifacts.upload.StageManager")
+@mock.patch("snowflake.cli.api.entities.utils.StageManager")
+@mock.patch("snowflake.cli._plugins.stage.diff.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.get_url")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.execute_query")
 @mock_streamlit_exists
 def test_deploy_streamlit_with_default_warehouse(
-    mock_execute_query, _, mock_stage_manager, __, temporary_directory
+    mock_execute_query, _, mock_stage_manager, __, ___, temporary_directory
 ):
     mock_stage_manager().get_standard_stage_prefix.return_value = "stage_root"
 
@@ -187,7 +191,8 @@ def test_deploy_streamlit_with_default_warehouse(
     )
 
 
-@mock.patch("snowflake.cli.api.artifacts.upload.StageManager")
+@mock.patch("snowflake.cli.api.entities.utils.StageManager")
+@mock.patch("snowflake.cli._plugins.stage.diff.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StageManager")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.get_url")
 @mock.patch("snowflake.cli._plugins.streamlit.manager.StreamlitManager.execute_query")
@@ -196,7 +201,7 @@ def test_deploy_streamlit_with_default_warehouse(
 )
 @mock_streamlit_exists
 def test_deploy_streamlit_with_grants(
-    mock_grants, _, __, mock_stage_manager, ___, temporary_directory
+    mock_grants, _, __, mock_stage_manager, ___, ____, temporary_directory
 ):
     mock_stage_manager().get_standard_stage_prefix.return_value = "stage_root"